Move FramedWebView to widgets/browser.py, clean up code imports
